Fetch Orders

+++ backend/src/main/java/edu/alexu/cse/dripmeup/dto/ItemDTO.java
@@ -13,7 +13,7 @@ public class ItemDTO {
     private String ProductName;
     private String ProductVariantSize;
     private String ProductVariantColor;
-    private Integer ProductVariantPrice;
+    private float ProductVariantPrice;
     private Integer ProductVariantQuantity;
-    private Integer ItemTotalPrice;
+    private float ItemTotalPrice;
 }
